rename file and types to match each other
