+++ src/DetailsView/components/overview-content/overview-help-section.tsx
@@ -4,16 +4,16 @@ import * as React from 'react';
 
 import { NamedSFC } from '../../../common/react/named-sfc';
 import { HyperlinkDefinition } from '../../../views/content/content-page';
-import { HelpLinks, HelpLinksDeps } from './overview-help-links';
+import { HelpLinks, HelpLinksDeps } from './help-links';
 
-export type HelpLinkDeps = HelpLinksDeps;
+export type OverviewHelpSectionDeps = HelpLinksDeps;
 
-export interface OverviewHelpProps {
-    deps: HelpLinkDeps;
+export interface OverviewHelpSectionProps {
+    deps: OverviewHelpSectionDeps;
     linkDataSource: HyperlinkDefinition[];
 }
 
-export const OverviewHelpSection = NamedSFC('OverviewHelpSection', (props: OverviewHelpProps) => {
+export const OverviewHelpSection = NamedSFC('OverviewHelpSection', (props: OverviewHelpSectionProps) => {
     return (
         <section className="overview-help-container">
             <h3 className="help-heading">Help</h3>
